Add base test class for scripted command tests to eliminate some duplicated code
